[ADD] Ajout du cas RA

+++ src/objects/Plane.java
@@ -7,25 +7,25 @@ package objects;
  */
 public class Plane {
 
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Plane(int x, int y) {
+	public Plane(double x, double y) {
 		super();
 		this.x = x;
 		this.y = y;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
